api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api"
authors = ["Joshua <jschmidtnj@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.3"
Flask-Cors = "^3.0.10"
flask-smorest = "^0.37.0"
loguru = "^0.6.0"
marshmallow = "^3.14.1"
peewee = "^3.14.9"
bidict = "^0.21.4"
aenum = "^3.1.8"
python-dotenv = "^0.19.2"
argon2-cffi = "^21.3.0"
PyJWT = "^2.3.0"
marshmallow-enum = {git = "https://github.com/h/marshmallow_enum.git"}
requests = "^2.27.1"
"ruamel.yaml" = "^0.17.21"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
pylint = "^2.12.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
